+++ CHANGES.md
@@ -2,6 +2,10 @@
 
 ## 2.0.0-SNAPSHOT
 
+*Sun Oct  6 16:32:01 EDT 2013*
+
+* Remove macros `cfn` and `defcfn` from public API.
+
 *Mon Sep 30 15:22:11 EDT 2013*
 
 * Remove `done` property from `Cell` type.